+++ src/context/index.ts
@@ -1,9 +1,4 @@
 export {
   useJobListingContext,
-  JobListingContext,
   JobListingProvider,
-  SearchFilterActions
 } from './jobListingContext';
-
-export type { SearchFilterAction } from './jobListingContext';
-
